refactor models into separate files
