+++ manager/models/functions.py
@@ -1,0 +1,13 @@
+from django.utils import timezone
+
+
+def today_midnight():
+    """
+    Return the time for midnight of the current day.
+
+    :returns: A datetime object for midnight of the current day.
+    """
+    midnight = timezone.now().replace(
+        hour=0, minute=0, second=0, microsecond=0
+    ) + timezone.timedelta(days=1)
+    return midnight
